JoinTree split up
Split up the various JoinTree implementations into separate files, improving the API
